convert the config files to a uci-defaults script
The uci-defaults scripts are run once on first boot and make the
necessary modifications.  This has the advantage of greatly simplifying
the FOOCAB.pl script and also of tracking upstream changes in the
OpenWrt configuration.  It is still subject to potential breakage, but
that was the case before.  Testing is needed when new revisions are
built, before deploying.

Signed-off-by: Russell Senior <russell@personaltelco.net>