trim some network configuration options for ipv6
network.globals provides a randomized site-local net block, which we do
not need and can lead to odd results in olsr6d route advertisements.  Also
ip6assign is appropriate only in cases where we have native ipv6 networks,
so remove it for the default case.  It should be added back, replacing the
ip6addr from the vpn, where native ipv6 is available.

Signed-off-by: Russell Senior <russell@personaltelco.net>