make uci-defaults script robust against emptiness
The /etc/uci-defaults/ptp.nocat.default script looks through
dhcp host records in uci and constructs pre-auth commands to append to
/usr/nocat/bin/initialize.fw.  However, if there are no dhcp host
records, a temporary file is not created and the script exits with a
non-zero return value, which prevents its removal from /etc/uci-defaults.

This change corrects that, allows the script to exit with 0 when there
are no dhcp host records.

The bug is mostly harmless.  The script will remain in /etc/uci-defaults
until someone adds a host record, and it has something to bite into.

Signed-off-by: Russell Senior <russell@personaltelco.net>