fixup for upstream change in mr24 base configuration
it used to be that single ethernet devices defaulted to doing
DHCP on that ethernet to facilitate initial configuration.
Our configuration relied on modifying that default configuration.
However, the MR24 in particular, and perhaps others changed the
way eth0 was configured with respect to DHCP. The result was
that our modification was not adequate to ensure that DHCP service
existed on the br-pub and br-priv networks. This change corrects
that.

Signed-off-by: Russell Senior <russell@personaltelco.net>